Refactor: rename techk folder by config
