+++ techk/apps/base/tests.py
@@ -1,3 +1,9 @@
-from django.test import TestCase
+# -*- coding: utf-8 -*-
+import pytest
+import requests
 
-# Create your tests here.
+
+def test_hello_world():
+    response = requests.get('http://localhost:8000')
+    dom = response.content
+    assert dom == b'Hello, world!'
