+++ techk/config/urls.py
@@ -15,7 +15,8 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 from django.contrib import admin
-from techk.views import index
+from apps.base.views import index
+
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
