+++ techk/pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-python_files = tests/*.py
+python_files = */tests.py
 addopts = -p no:warnings --pdb
